% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{benchmark_boxplot}
\alias{benchmark_boxplot}
\title{Normal box plot}
\usage{
benchmark_boxplot(benchmark_result, log = FALSE,
  xlab = "Docker Images", ylab = "Duration (seconds)", ylim = NULL,
  ...)
}
\arguments{
\item{benchmark_result}{Benchmark result from altRnative}

\item{log}{Set to TRUE If use log (time), else FALSE}

\item{xlab}{The lab for x axis}

\item{ylab}{The lab for y axis}

\item{ylim}{Range for ye axis, if NULL, set to max + 1 seconds}

\item{...}{Parameters passed on boxplot method}
}
\description{
Plot a box plot from the result of benchmarking with some default
}
