% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmarks}
\alias{benchmarks}
\title{Internal function for benchmark execution}
\usage{
benchmarks(run_function, code_or_file, platforms = c("debian", "ubuntu"),
  r_implementations = c("gnu-r", "mro"), volumes = volumes,
  times = 3, pull_image = pull_image, ...)
}
\arguments{
\item{platforms}{List of platforms}

\item{r_implementations}{List of R implementations}

\item{volumes}{Volume mapping from host to container. Passed to \link{run_code}}

\item{times}{How many times the code will be run. Passed to \link{microbenchmark}}

\item{pull_image}{If set to TRUE, the needed docker image will be pulled first}

\item{...}{Parameters for \link[microbenchmark]{microbenchmark}}
}
\description{
Internal function for benchmark execution
}
\keyword{internal}
