% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmarks_code}
\alias{benchmarks_code}
\title{Run benchmark for R code across R implementation and platform}
\usage{
benchmarks_code(code, platforms = c("debian", "ubuntu"),
  r_implementations = c("gnu-r", "mro"), ...)
}
\arguments{
\item{code}{An expression or string of R code}

\item{platforms}{List of platform}

\item{r_implementations}{List of R implementation}

\item{...}{Parameters for \link{microbenchmark}}
}
\description{
Run benchmark for R code across R implementation and platform.
}
\examples{
benchmarks_code(code = "1 + 1", times = 3)
}
