% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmarks_file}
\alias{benchmarks_file}
\title{Run benchmark for R code file across R implementation and platform}
\usage{
benchmarks_file(r_file, platforms = c("debian", "ubuntu"),
  r_implementations = c("gnu-r", "mro"), ...)
}
\arguments{
\item{r_file}{A file of R code}

\item{platforms}{List of platform}

\item{r_implementations}{List of R implementation}

\item{...}{Parameters for \link{microbenchmark}}
}
\description{
Run benchmark for R code file across R implementation and platform.
}
\examples{
file_path <- system.file('extdata/test.R', package = 'altRnative')
benchmarks_file(file_path, times = 3)
}
