% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compatibility_table}
\alias{compatibility_table}
\title{Return a compatibility table}
\usage{
compatibility_table()
}
\description{
Return a table that consist of platfom, R implementation, docker image name.
}
\examples{
compatibility_table()
}
