% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{docker_run_code}
\alias{docker_run_code}
\title{Run a code in a docker image}
\usage{
docker_run_code(code, docker_image)
}
\arguments{
\item{code}{An expression or string of R code}

\item{docker_image}{A docker image name}
}
\description{
Run a code as a string in a docker image
}
\examples{
library("altRnative")

# With string
docker_run_code("a = 1 + 1", "ismailsunni/gnur-3.6.1-debian-geospatial")

# With expression
docker_run_code(expression(1 + 1), "ismailsunni/gnur-3.6.1-debian-geospatial")
docker_run_code(expression(a = 1 + 1), "ismailsunni/gnur-3.6.1-debian-geospatial")
docker_run_code(expression(install.packages("ctv")), "ismailsunni/gnur-3.6.1-debian-geospatial")

# With multiple string
docker_run_code(c("a = 1 + 1", "b = a + 2", "print(b)"), "ismailsunni/gnur-3.6.1-debian-geospatial")
code = c("install.packages('ctv')", "library('ctv')", "available.views()")
docker_run_code(code, "ismailsunni/gnur-3.6.1-debian-geospatial")

# With multiple expressions
# This one is not working, see https://github.com/ismailsunni/altRnative/issues/1
# docker_run_code(c(expression(a = 1 + 1, b = a + 2)), "ismailsunni/gnur-3.6.1-debian-geospatial")
# This one is working
code = expression(install.packages("ctv"), library("ctv"), available.views())
docker_run_code(code, "ismailsunni/gnur-3.6.1-debian-geospatial")
}
