% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{normalize_benchmark_boxplot}
\alias{normalize_benchmark_boxplot}
\title{Normalize box plot}
\usage{
normalize_benchmark_boxplot(benchmark_result, log = FALSE,
  xlab = "Docker Images", ylab = "Duration Ratio", ylim = NULL, ...)
}
\description{
Plot a normalize box plot from the result of benchmarking with some default values
}
