% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_docker_image}
\alias{pull_docker_image}
\title{Pull docker image from Docker Hub}
\usage{
pull_docker_image(platforms = c("debian", "ubuntu"),
  r_implementations = c("gnu-r", "mro"), stevedore_opts = list())
}
\arguments{
\item{platforms}{List of platform}

\item{r_implementations}{List of R implementation}

\item{stevedore_opts}{options passed to \link[stevedore]{docker_client}}
}
\value{
the output of \link{stevedore}'s function \code{docker$image$pull}, or \code{NULL} if no image was found
}
\description{
Works for public images only and checks if platform and R are supported, see \link{compatibility_table}¸
}
\details{
For supported docker images only.
}
\examples{
\dontrun{
pull_docker_image('debian', 'gnu-r')
}

}
