% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run a code in a docker image}
\usage{
run(code, docker_image)
}
\arguments{
\item{code}{An expression of R code}

\item{docker_image}{A docker image name}
}
\description{
Run a a code as a string in a docker image
}
\examples{
library("altRnative")
run(expression(1 + 1), "ismailsunni/gnur-3.6.1-debian-geospatial")
run(expression(install.packages("ctv")), "ismailsunni/gnur-3.6.1-debian-geospatial")
run(c(expression(install.packages("ctv")), expression(library("ctv")), expression(available_views())), "ismailsunni/gnur-3.6.1-debian-geospatial")
}
