% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_code}
\alias{run_code}
\title{Run a R file in a docker image  of a platform and an R implementation}
\usage{
run_code(code, platform = "debian", r_implementation = "gnu-r")
}
\arguments{
\item{code}{An expression or string of R code}

\item{platform}{The platform name see \link{supported_platforms}}

\item{r_implementation}{The R implementation name. See \link{supported_Rs}}
}
\description{
Run a R file in a docker image  of a platform and an R implementation
}
\examples{
run_code("a = 1 + 1", 'debian', 'gnu-r')
}
