% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{supported_platforms}
\alias{supported_platforms}
\title{Return supported platforms}
\usage{
supported_platforms()
}
\description{
Return list of supported platforms with their key to access.
}
\examples{
supported_platforms()
}
