% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{supported_Rs}
\alias{supported_Rs}
\title{Return supported R implementations}
\usage{
supported_Rs()
}
\description{
Return list of supported R implementations with their key to access.
}
\examples{
supported_Rs()
}
