% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{benchmark_boxplot}
\alias{benchmark_boxplot}
\title{Normal box plot}
\usage{
benchmark_boxplot(benchmark_result, log = FALSE,
  xlab = "Docker Images", ylab = "Duration (seconds)", ylim = NULL,
  ...)
}
\description{
Plot a box plot from the result of benchmarking with some default
}
