% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmarks_file}
\alias{benchmarks_file}
\title{Run benchmark for R code file across R implementation and platform}
\usage{
benchmarks_file(r_file, platforms = c("debian", "ubuntu"),
  r_implementations = c("gnu-r", "mro"), volumes = NULL,
  pull_image = FALSE, times = 3, ...)
}
\arguments{
\item{r_file}{A file of R code}

\item{platforms}{List of platforms}

\item{r_implementations}{List of R implementations}

\item{volumes}{Volume mapping from host to container. Passed to \link{run_code}}

\item{pull_image}{If set to TRUE, the needed docker image will be pulled first}

\item{times}{How many times the code will be run. Passed to \link{microbenchmark}}

\item{...}{Parameters for \link[microbenchmark]{microbenchmark}}
}
\description{
Run benchmark for R code file across R implementation and platform.
}
\examples{
\dontrun{
file_path <- system.file('extdata/test.R', package = 'altRnative')
benchmarks_file(file_path, times = 3)
}
}
