% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmarks_file}
\alias{benchmarks_file}
\title{Run benchmark for R code file across R implementation and platform}
\usage{
benchmarks_file(r_file, r_implementations = c("gnu-r", "mro"),
  platforms = c("debian", "ubuntu"), ...)
}
\arguments{
\item{r_file}{A file of R code}

\item{r_implementations}{List of R implementation}

\item{platforms}{List of platform}
}
\description{
Run benchmark for R code file across R implementation and platform.
}
\examples{
file_path <- system.file('extdata/test.R', package = 'altRnative')
benchmarks_file(file_path)
}
