% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{normalize_benchmark_boxplot}
\alias{normalize_benchmark_boxplot}
\title{Normalize box plot}
\usage{
normalize_benchmark_boxplot(benchmark_result, log = FALSE,
  xlab = "Docker Images", ylab = "Duration Ratio", ylim = NULL, ...)
}
\arguments{
\item{benchmark_result}{Benchmark result from altRnative}

\item{log}{Set to TRUE If use log (time), else FALSE}

\item{xlab}{The lab for x axis}

\item{ylab}{The lab for y axis}

\item{ylim}{Range for ye axis, if NULL, set to max + 1 seconds}

\item{...}{Parameters passed on boxplot method}
}
\description{
Plot a normalize box plot from the result of benchmarking with some default values
}
