% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_docker_image}
\alias{pull_docker_image}
\title{Pull docker image from docker hub. Public image only.}
\usage{
pull_docker_image(
  platforms = c("debian", "ubuntu"),
  r_implementations = c("gnu-r", "mro")
)
}
\arguments{
\item{platforms}{List of platform}

\item{r_implementations}{List of R implementation}
}
\description{
For supported docker images only.
}
\examples{
\dontrun{
pull_docker_image('debian', 'gnu-r')
}

}
