% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run a code in a docker image}
\usage{
run(code, docker_image)
}
\arguments{
\item{code}{An expression or string of R code}

\item{docker_image}{A docker image name}
}
\description{
Run a a code as a string in a docker image
}
\examples{
library("altRnative")

# With string
run("a = 1 + 1", "ismailsunni/gnur-3.6.1-debian-geospatial")

# With expression
run(expression(1 + 1), "ismailsunni/gnur-3.6.1-debian-geospatial")
run(expression(a = 1 + 1), "ismailsunni/gnur-3.6.1-debian-geospatial")
run(expression(install.packages("ctv")), "ismailsunni/gnur-3.6.1-debian-geospatial")

# With multiple string
run(c("a = 1 + 1", "b = a + 2", "print(b)"), "ismailsunni/gnur-3.6.1-debian-geospatial")
run(c("install.packages('ctv')", "library('ctv')", "available.views()"), "ismailsunni/gnur-3.6.1-debian-geospatial")

# With multiple expressions
# This one is not working, see https://github.com/ismailsunni/altRnative/issues/1
run(c(expression(a = 1 + 1, b = a + 2)), "ismailsunni/gnur-3.6.1-debian-geospatial")
# This one is working
run(c(expression(install.packages("ctv")), expression(library("ctv")), expression(available.views())), "ismailsunni/gnur-3.6.1-debian-geospatial")
}
