% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_file}
\alias{run_file}
\title{Run a R file in a docker image}
\usage{
run_file(r_file, docker_image)
}
\arguments{
\item{r_file}{A file of R code}

\item{docker_image}{A docker image name}
}
\description{
Run a R file in a docker image
}
\examples{
library("altRnative")
run_file("check_packages.R", "ismailsunni/gnur-3.6.1-debian-geospatial")
}
